+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Issue #314: Simplify yaml file schema
 - Issue #316: Updating polar map takes too much cpu
+- Issue #320: Differentiate contact updates in the radar map
 
 ## [0.12.1] 2024-01-14